New internal EmailData class; use ImmutableList
