+++ src/main/java/desi/juan/email/api/EmailBody.java
@@ -25,79 +25,40 @@
  */
 package desi.juan.email.api;
 
-import javax.mail.internet.ContentType;
-import javax.mail.internet.MimeUtility;
-import java.io.UnsupportedEncodingException;
 import java.nio.charset.Charset;
 
-import static desi.juan.email.api.EmailConstants.DEFAULT_CHARSET;
 import static desi.juan.email.api.EmailConstants.DEFAULT_CONTENT_TYPE;
-import static java.lang.String.format;
 
 /**
- * Represents and enables the construction of the content of an email with a body of type "text/*" and a specific character
- * encoding.
+ * Represents and enables the construction of the content of an email.
+ *
+ * @see EmailData
  */
-public class EmailBody {
+public class EmailBody extends EmailData {
 
   /**
-   * Text body of the message content.
+   * @return the body content of the {@link Email} as a {@link String} value.
    */
-  private final String content;
-
-  /**
-   * @return the body content of the email as a {@link String} value.
-   */
-
+  @Override
   public String getContent() {
-    return content;
+    return (String) super.getContent();
   }
 
   /**
-   * Format of the body text. Example: "text/html" or "text/plain".
-   */
-  private final String format;
-
-  /**
-   * @return format of the body text. Example: "text/html" or "text/plain".
-   */
-  public String getFormat() {
-    return format;
-  }
-
-  /**
-   * The charset of the body text
-   */
-  private final Charset charset;
-
-  /**
-   * @return the charset of the body text
-   */
-  public Charset getCharset() {
-    return charset;
-  }
-
-  /**
-   * @return the content type of the body, includes email format and charset
-   */
-  public String getContentType() {
-    return format("%s; charset=%s", format, charset);
-  }
-
-  /**
-   * EmailBody constructor
-   *
    * @param content
    * @param charset
-   * @param emailFormat
+   * @param format
    */
-  public EmailBody(String content, Charset charset, String emailFormat) {
-    if (content == null || charset == null || emailFormat == null) {
-      throw new IllegalArgumentException("No param can be null");
-    }
-    this.content = content;
-    this.charset = charset;
-    this.format = emailFormat.toLowerCase();
+  public EmailBody(String content, Charset charset, String format) {
+    super(content, charset, format);
+  }
+
+  /**
+   * @param content
+   * @param contentType
+   */
+  public EmailBody(String content, String contentType) {
+    super(content, contentType);
   }
 
   /**
@@ -106,7 +67,7 @@ public class EmailBody {
    * @see #EmailBody(String, String)
    */
   public EmailBody(String content) {
-    this(content, DEFAULT_CONTENT_TYPE);
+    super(content, DEFAULT_CONTENT_TYPE);
   }
 
   /**
@@ -116,58 +77,5 @@ public class EmailBody {
    */
   public EmailBody() {
     this("");
-  }
-
-  /**
-   * @param content
-   * @param contentType
-   * @see #EmailBody(String, ContentType)
-   */
-  public EmailBody(String content, String contentType) {
-    this(content, toContentType(contentType));
-  }
-
-  /**
-   * @param content
-   * @param contentType
-   * @see #EmailBody(String, Charset, String)
-   */
-  protected EmailBody(String content, ContentType contentType) {
-    this(content, getCharset(contentType), getFormat(contentType));
-  }
-
-  /**
-   * @param contentType
-   * @return
-   */
-  private static Charset getCharset(ContentType contentType) {
-    String csParam = contentType.getParameter("charset");
-    try {
-      String csStr = MimeUtility.decodeText(csParam);
-      return Charset.forName(csStr);
-    } catch (UnsupportedEncodingException e) {
-      return DEFAULT_CHARSET;
-    }
-  }
-
-  /**
-   * @param contentType
-   * @return
-   */
-  private static ContentType toContentType(String contentType) {
-    try {
-      return new ContentType(contentType);
-    } catch (Exception e) {
-      // TODO: should this do something else?
-      return new ContentType();
-    }
-  }
-
-  /**
-   * @param contentType
-   * @return
-   */
-  private static String getFormat(ContentType contentType) {
-    return contentType.getBaseType();
   }
 }
